+++ osidb/filters.py
@@ -222,6 +222,10 @@ class FlawFilter(DistinctFilterSet):
             + LTE_GTE_LOOKUP_EXPRS
             + DATE_LOOKUP_EXPRS,
             "affects__trackers__errata__advisory_name": ["exact"],
+            "affects__trackers__errata__shipped_dt": ["exact"]
+            + LT_GT_LOOKUP_EXPRS
+            + LTE_GTE_LOOKUP_EXPRS
+            + DATE_LOOKUP_EXPRS,
             # Acknowledgment fields
             "acknowledgments__uuid": ["exact"],
             "acknowledgments__name": ["exact"],